+++ MaterialDesignThemes.Wpf/Converters/CircularProgressBar/ArcSizeConverter.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+
+namespace MaterialDesignThemes.Wpf.Converters.CircularProgressBar
+{
+    public class ArcSizeConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is double && ((double)value > 0.0))
+            {
+                return new Size((double)value / 2, (double)value / 2);
+            }
+
+            return new Point();
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return Binding.DoNothing;
+        }
+    }
+}
