+++ MaterialDesignThemes.Wpf/Converters/CircularProgressBar/StartPointConverter.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+
+namespace MaterialDesignThemes.Wpf.Converters.CircularProgressBar
+{
+    public class StartPointConverter : IValueConverter
+    {
+        [Obsolete]
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is double && ((double) value > 0.0))
+            {
+                return new Point((double)value / 2, 0);
+            }
+
+            return new Point();
+        }
+
+        [Obsolete]
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return Binding.DoNothing;
+        }
+
+    }
+}
