+++ MaterialDesignThemes.Wpf/MaterialDesignThemes.Wpf.csproj
@@ -197,6 +197,13 @@
     <Compile Include="ColorZone.cs" />
     <Compile Include="Converters\BrushToRadialGradientBrushConverter.cs" />
     <Compile Include="Converters\CalendarDateCoalesceConverter.cs" />
+    <Compile Include="Converters\CircularProgressBar\ArcEndPointConverter.cs" />
+    <Compile Include="Converters\CircularProgressBar\ArcSizeConverter.cs" />
+    <Compile Include="Converters\CircularProgressBar\LocalEx.cs" />
+    <Compile Include="Converters\CircularProgressBar\LargeArcConverter.cs" />
+    <Compile Include="Converters\CircularProgressBar\RotateTransformCentreConverter.cs" />
+    <Compile Include="Converters\CircularProgressBar\RotateTransformConverter.cs" />
+    <Compile Include="Converters\CircularProgressBar\StartPointConverter.cs" />
     <Compile Include="Converters\ClockItemIsCheckedConverter.cs" />
     <Compile Include="Converters\ClockLineConverter.cs" />
     <Compile Include="Converters\NotConverter.cs" />
